work push

+++ src/render/shape/uniform.rs
@@ -1,15 +1,15 @@
 #[repr(C)]
 #[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable, Default)]
 pub struct WindowUniform {
-    pub width: u32,
-    pub height: u32,
+    pub width: f32,
+    pub height: f32,
 }
 
 impl WindowUniform {
     pub fn new() -> Self {
         Self {
-            width: 0,
-            height: 0,
+            width: 0.0,
+            height: 0.0,
         }
     }
 }
